refactor: fix circular dependency, remove unused variables, join same condition result into one
