+++ src/components/TextField/variants.ts
@@ -1,4 +1,20 @@
-import { ReadOnlyVariant, IconVariant, InputVariant } from './TextField';
+export enum InputVariant {
+  Default = 'default',
+  Error = 'error',
+  Disabled = 'disabled',
+  ReadonlyInfo = 'readonly-info',
+  ReadonlyConfirm = 'readonly-confirm',
+}
+
+export enum ReadOnlyVariant {
+  ReadonlyInfo = 'readonly-info',
+  ReadonlyConfirm = 'readonly-confirm',
+}
+
+export enum IconVariant {
+  None = 'none',
+  Error = 'error',
+}
 
 export const getVariants = (
   readOnly?: boolean | ReadOnlyVariant,
@@ -10,12 +26,7 @@ export const getVariants = (
       variant: InputVariant.Disabled,
       iconVariant: IconVariant.None,
     };
-  } else if (readOnly === true) {
-    return {
-      variant: InputVariant.ReadonlyInfo,
-      iconVariant: IconVariant.None,
-    };
-  } else if (readOnly === ReadOnlyVariant.ReadonlyInfo) {
+  } else if (readOnly === true || readOnly === ReadOnlyVariant.ReadonlyInfo) {
     return {
       variant: InputVariant.ReadonlyInfo,
       iconVariant: IconVariant.None,
